+++ IdleSlayerMods.Common/Extensions/LoggerExtension.cs
@@ -6,6 +6,7 @@ public static class LoggerExtension
 {
     public static void Debug(this MelonLogger.Instance logger, string message)
     {
-        logger.MsgPastel(ConsoleColor.DarkGray, message);
+        if (!ModUtils.DebugMode) return;
+        logger.MsgPastel(ConsoleColor.DarkGray, $"[DEBUG] {message}");
     }
 }